+++ creator-node/src/utils/contentNodeUtils.ts
